Create vary_order merits_task and proceeding_type_merits_task
Aim is to rename reason_for_new_application --> vary_order
as from discussion between JS and CB it was agreed that
the name of the merits task is more clearly and concisely
covered by vary_order.

This name is to be used by Apply as well for the merits
question logic.