Cycle git-crypt symmetric key
As a result of the [Circle CI incident][incident], we need to rotate our secrets
stored in Circle CI.

One of the secrets stored is `GIT_CRYPT_KEY`, which can be used to unlock
encrypted files stored in the repository.

Since we use `git-crypt` to store our secrets, this key needs to be rotated as a
priority.

This replaces the encrypted files with new versions that have been encrypted
with the new symmetric key.

[incident]: https://circleci.com/blog/january-4-2023-security-alert/