Chore: Update circle flow
This aims to update the flow to be similar for all services.

If both deploys succeed, it sends the message to slack and
allows the hold to be released. This means that when deploying
in a hurry, emergency fixes, etc; and you are waiting to deploy
you are not blocked waiting for the the slack message to be sent.

It also means that if slack is offline for any reason it does not block deploys